P2P-reachable unbounded Vec::with_capacity on attacker-chosen counts (witness
stack, tx count, and friends) let a tiny message pre-reserve hundreds of MB
before a byte was read. Clamped to MAX_VECTOR_ALLOCATE via the length-guarded
decode_vec helper. Verified: 126 tests green across qubitcoin-consensus and
qubitcoin-serialize.

+++ crates/qubitcoin-consensus/src/check.rs
@@ -68,8 +68,11 @@ pub fn check_transaction(tx: &Transaction, state: &mut TxValidationState) -> boo
         // locktime
         tx.lock_time.encode(&mut no_witness_buf).unwrap();
 
-        let stripped_size = no_witness_buf.len() as u32;
-        if stripped_size * WITNESS_SCALE_FACTOR > MAX_BLOCK_WEIGHT {
+        // Compute in u64 to avoid a u32 multiply overflow: a sufficiently large
+        // stripped_size (~1 GiB) would wrap `stripped_size * WITNESS_SCALE_FACTOR`
+        // and could let an oversize transaction slip past this check.
+        let stripped_size = no_witness_buf.len() as u64;
+        if stripped_size * WITNESS_SCALE_FACTOR as u64 > MAX_BLOCK_WEIGHT as u64 {
             return state.invalid(TxValidationResult::Consensus, "bad-txns-oversize", "");
         }
     }
